test(application): fixture 분리
